Rename parser extension classes
