+++ app/src/org/commcare/activities/FormEntryActivity.java
@@ -55,8 +55,8 @@ import org.commcare.activities.components.ImageCaptureProcessing;
 import org.commcare.dalvik.BuildConfig;
 import org.commcare.dalvik.R;
 import org.commcare.views.media.MediaLayout;
-import org.odk.collect.android.jr.extensions.IntentCallout;
-import org.odk.collect.android.jr.extensions.PollSensorAction;
+import org.commcare.android.javarosa.IntentCallout;
+import org.commcare.android.javarosa.PollSensorAction;
 import org.commcare.interfaces.AdvanceToNextListener;
 import org.commcare.interfaces.FormSaveCallback;
 import org.commcare.interfaces.FormSavedListener;
@@ -68,7 +68,6 @@ import org.commcare.logging.analytics.TimedStatsTracker;
 import org.commcare.logic.FormController;
 import org.commcare.logic.PropertyManager;
 import org.commcare.models.ODKStorage;
-import org.commcare.models.database.DbUtil;
 import org.commcare.preferences.FormEntryPreferences;
 import org.commcare.provider.FormsProviderAPI.FormsColumns;
 import org.commcare.provider.InstanceProviderAPI;
